requirements.txt
mysql-connector-python==8.1.0
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn==0.23.1
python-multipart